Turned the WP7 folder into a git submodule.
The dedicated repository is https://git.code.sf.net/p/ddmorewp7/code.

In the new layout of this repo, each WP will be a submodule, thus making it
accessible without needing to checkout the entire DDMoRe repository. It is
advisable that the WP repositories are also partitioned into submodules
dedicated to each individual task.