Ticket #80: export of target code blocks added - target code blocks are converted to an xml format and stored in the folder target: target/model_name.xml.
There was a problem to include snowyaml.jar and we do not really need
it, it is easy to parse the simple .xml file. The choice of XML over
JSON like syntax is to be consistent with the generated pharmML. Also
target code blocks can contain empty lines, figure brackets etc., so it
may be simpler to parse XML. 

UI module updated to refresh nested output folders.