Added basic support for library description. Libraries are called as ordinary functions which return a set of predefined variables. These variables are valiadted by the MDL-IDE and cannot be changed by the user. For example, 'nmadvan' library returnes variables {A, F}, where A is of type TYPE_VECTOR_REAL, and type of F is probably TYPE_REAL (all types are undefined for now as we don't do type checking yet). 'PK' library returns variables {V, KL, KA} for now.
The attribute 'output' is not needed anymore and will not be any longer
validated, so update the LIBRARY definitions of existing models by
removing it!

The sets of returned variable sets will depend on library input
parameters. Rigth combinations of parameters and the corresponding
returned variables is yet to be specified!