DDMORE-1467: BlkStmtFactory.fromJSON() can now create from single item.
R cannot distinguish between a string and a character vector of length
1, hence a variable-name list (vector when read in to R) containing one
variable gets translated to JSON as a string rather than a list of
strings containing one element. So if BlockStatementFactory.fromJSON()
is passed a String as a Value in the incoming Map containing the JSON,
rather than a List of String, then wrap this String in a List before
passing it on.