DDMORE-1506: Refactoring to circumvent org.eclipse.core.runtime issue.
The issue being:
MDLToPharmMLConverterTest.testPerformConvertForValidMdlFile() is
suddenly failing with classloader error: class
"org.eclipse.core.runtime.IPath"'s signer information does not match
signer information of other classes in the same package.

Analysis: MdlStandaloneSetup.createInjectorAndDoEMFRegistration()
initialisation loads certain org.eclipse.core.runtime.* classes via the
Platform.isRunning() call in EMFPlugin. This caches the certificates for
that problematic package. There is a org.ddmore.mdl.validation.Utils
class that is used to get the list of objects in a given MOG, and this
is invoked from MdlToPharmMLConverter. This Utils class has a reference
to org.eclipse.core.runtime.Path and it is the classloading of this
class from the MDL->PharmML converter rather than from within the Xtext
initialisation that triggers the error (this class isn't already loaded
as part of the above mentioned initialisation).

Quick(ish) fix solution implemented by this commit: Refactor to move the
reference to that Path class into PropertyValidator (where it is
ultimately used) and refactor out code from that PropertyValidator class
that is referenced (directly or indirectly) from the
MDL2PharmMLConverter rather than being referenced from the running IDE
or from within the Xtext code.