DDMORE-1550: Don't promote tool-specific compatibility warnings to errs
The introduction of NONMEM- and MONOLIX-specific compatibility warnings,
using the same "eu.ddmore.mdl.validation.unsupported.feature" prefix but
with ".nonmem" and ".monolix" suffixes, means that my existing check
that an Issue's code string starts with the
"eu.ddmore.mdl.validation.unsupported.feature" in order to promote a
warning to an error, is no longer valid, since the tool-specific
compatibility warnings should *not* be promoted to errors. So check for
equality of "eu.ddmore.mdl.validation.unsupported.feature" string
instead. With the caveat that uses of any introduced "sub-classes" of
this string will need to explicitly be promoted to errors if required
now.